Separate conversation component and simplify logic
