+++ translations/all-messages.json
@@ -1,5 +1,10 @@
 {
   "conversation-list-item.unnamedUser": "Unnamed User",
+  "Messages.newMessagePlaceholder": "Type something...",
+  "purchase-summary.purchasedFrom": "Purchased from {sellerLink}",
+  "purchase-summary.soldTo": "Sold to {buyerLink}",
+  "purchase-summary.purchasedFromOn": "Purchased from {sellerName} on {date}",
+  "purchase-summary.soldToOn": "Sold to {buyerName} on {date}",
   "connectivity.mainEthereumNetwork": "Main Ethereum Network",
   "connectivity.mordenTestNetwork": "Morden Test Network",
   "connectivity.ropstenTestNetwork": "Ropsten Test Network",
@@ -114,11 +119,6 @@
   "MessageNew.send": "Send",
   "MessageNew.encryptionNotice": "Your message will be encrypted. It will only be visible to you, the recipient, and an arbitrator in the event that a dispute arises.",
   "message.enable": "Enable Messaging",
-  "Messages.newMessagePlaceholder": "Type something...",
-  "purchase-summary.purchasedFrom": "Purchased from {sellerLink}",
-  "purchase-summary.soldTo": "Sold to {buyerLink}",
-  "purchase-summary.purchasedFromOn": "Purchased from {sellerName} on {date}",
-  "purchase-summary.soldToOn": "Sold to {buyerName} on {date}",
   "messaging-provider.congrats": "Congratulations! You can now message other users across the Origin Platform ecosystem! Why not start by taking a look around and telling us what you think about our DApp?",
   "messaging-provider.unnamedUser": "Unnamed User",
   "messaging-provider.welcome": "Welcome to the Origin Protocol Demo DApp! Our innovative messaging platform allows you to communicate with other Origin users in a secure and decentralized way. All of your messages will be encrypted and stored on IPFS. Messages can only be read by you, the recipient, and potentially a third-party arbitrator if either of you initiates a dispute and grants access to a cryptographically-generated, shared, private key.\n\nIn order to enable messaging you will need to use your Ethereum wallet to sign two statements. This will not cost any ETH, Origin Token, or gas.",
